+++ core-ng-search/src/main/java/core/framework/search/impl/ElasticSearchLogInterceptor.java
@@ -7,11 +7,13 @@ import org.apache.http.HttpRequest;
 import org.apache.http.HttpRequestInterceptor;
 import org.apache.http.RequestLine;
 import org.apache.http.protocol.HttpContext;
-import org.apache.http.util.EntityUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.Reader;
+import java.nio.charset.StandardCharsets;
 import java.util.Set;
 
 /**
@@ -33,13 +35,26 @@ public class ElasticSearchLogInterceptor implements HttpRequestInterceptor {
         }
     }
 
-    private record BodyParam(HttpEntity entity) implements LogParam {
+    record BodyParam(HttpEntity entity) implements LogParam {
         @Override
         public void append(StringBuilder builder, Set<String> maskedFields, int maxParamLength) {
-            try {
-                builder.append(EntityUtils.toString(entity));
+            // refer to co.elastic.clients.transport.rest_client.RestClientTransport.prepareLowLevelRequest
+            // it always uses ByteArrayEntity, thus always has content length
+            try (Reader reader = new InputStreamReader(entity.getContent(), StandardCharsets.UTF_8)) {
+                long length = entity.getContentLength();
+                boolean truncate = false;
+                if (length > maxParamLength) {
+                    length = maxParamLength;
+                    truncate = true;
+                }
+                char[] buffer = new char[(int) length];
+                int read = reader.read(buffer);
+                builder.append(buffer, 0, read);
+                if (truncate) {
+                    builder.append("...(truncated)");
+                }
             } catch (IOException e) {
-                throw new Error(e);
+                throw new Error(e); // not expected io exception, as it's from ByteArrayEntity
             }
         }
     }